Add print functionality to entry view
When the user clicks the print button in the entry view, the preview
body is cloneNode'd to a new div#printpreview. The annotations are
rendered in an <ol> and appended to the div#printpreview. The
div#printpreview is appended to the <body>. We will be adding
information to the printpreview like the entry name and current date.