Fix bug where clicking on annotation would not show editor
The EntryPreview would subscribe to an add annotation event in initialize.
But because the view was not destroyed (cached by viewManager), the subscribe
would attach multiple times. Removed subscribe and used triggers instead, but
this should ultimately be fixed in the viewManager. For example by freezing
and unfreezing views (temporarily remove listeners and re-attach listeners).