Clean up working of edit multiple entry metadata
Added Publishable checkbox. It was not possible to empty a value, so I
added some checkboxes; when checked the value will be empty ('') in case
of a text input or false in case of a checkbox. The toggling of the form
was done using a class which set the display to block, but then I had to
reset the form, so now I just rerender the form everytime it is to be
shown. This is not much overhead and much cleaner. The toggling of the
save button is now done with a settings hash. If the hash is populated
(not empty) than the save button is to be activated. This removed the
need for an extra loop when saving, because the settins hash can be
passed directly instead of getting the data from the DOM.