Highlight full text search queryies in entry preview
Biggest part of making highlighting for full text search queries working
was changing the rendering of the search results. Instead of passing the
faceted search responseModel to a template and looping the results, we
now loop the results in the view and attach rendered entry listitem
views to a documentfrag. This way we can pass data easily between the
search view and the entry view. Along the way, I noticed the data
passing was unnecessary, but still it works cleaner (no more data-id
attributes).